miniproject01/pyproject.toml
[project]
name = "miniproject01"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "datasets>=4.5.0",
    "ipywidgets>=8.1.8",
    "langchain>=1.2.7",
    "langchain-core>=1.2.7",
    "langchain-groq>=1.1.1",
    "langchain-text-splitters>=1.1.0",
    "pandas>=3.0.0",
    "pypdf>=6.6.2",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "rouge-score>=0.1.2",
    "sentence-transformers>=5.2.2",
    "tqdm>=4.67.1",
    "weaviate-client>=4.19.2",
]
